Reconfigured to include a new Delegates:  MaturePresent.  Removed SiteVars.Cohorts.  Changed InitCommunityPixel to int from byte.
git-svn-id: http://Marc-PC/svn/Full@139 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/succession-library/trunk/src/InitCommPixel.cs
@@ -7,7 +7,7 @@ namespace Landis.Library.Succession
 {
     public class InitCommPixel : Pixel
     {
-        public Band<byte> MapCode  = "The numeric code for each initial community";
+        public Band<int> MapCode  = "The numeric code for each initial community";
 
         public InitCommPixel() 
         {
